requirements.txt
pytest~=7.4.4
Appium-Python-Client~=3.1.1